Updating scripting to pull latest JSON file from github and check for conditions before running.
There is a check for checking if the current user is sudo.  Typically if
this is deployed through an MDM the user will be running as sudo, but
for testing the user needs to be elevated.

Script first checks if Nudge is installed on the machine, then checks if
Nudge is already running.  Want to avoid multiple instances of Nudge
running.

Curl the latest JSON from github and then swap the temporary file into
the correct location.